# openapi.yaml
---
components:
  schemas:
    Base:
      properties:
        id:
          type: integer
        name:
          type: string
      type: object
    Order:
      properties:
        id:
          type: integer
        total_price:
          type: number
        poke:
          items:
            properties:
              id:
                type: integer
              base_id:
                type: integer
              protein_ids:
                items:
                  type: integer
                type: array
              ingredient_ids:
                items:
                  type: integer
                type: array
              price:
                type: number
              portion_id:
                type: integer
            type: object
          type: array
      type: object
    Ingredient:
      properties:
        id:
          type: integer
        name:
          type: string
      type: object
    CreateOrderRequest:
      properties:
        poke_ids:
          items:
            type: integer
          type: array
        total_price:
          type: integer
      type: object
    CreateOrderResponse:
      properties:
        id:
          type: integer
      type: object
    CreatePokeRequest:
      properties:
        base_id:
          type: integer
        price:
          type: integer
        ingredient_ids:
          items:
            type: integer
          type: array
        portion_id:
          type: integer
        protein_ids:
          items:
            type: integer
          type: array
      type: object
    CreatePokeResponse:
      properties:
        id:
          type: integer
      type: object
    Portion:
      properties:
        id:
          type: integer
        name:
          type: string
      type: object
    Protein:
      properties:
        id:
          type: integer
        name:
          type: string
      type: object
    UpdatePokeRequest:
      properties:
        base_id:
          type: integer
        price:
          type: integer
        ingredient_ids:
          items:
            type: integer
          type: array
        portion_id:
          type: integer
        protein_ids:
          items:
            type: integer
          type: array
      type: object
info:
  description: API for managing poke bowl orders.
  title: Poke Bowl API
  version: 1.0.0
openapi: 3.0.0
paths:
  /assets/bases:
    get:
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Base'
                type: array
          description: A list of bases
      summary: Get list of bases
      tags:
        - assets
  /assets/ingredients:
    get:
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Ingredient'
                type: array
          description: A list of ingredients
      summary: Get list of ingredients
      tags:
        - assets
  
  /assets/portions:
    get:
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Portion'
                type: array
          description: A list of portions
      summary: Get list of portions
      tags:
        - assets
  /assets/proteins:
    get:
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Protein'
                type: array
          description: A list of proteins
      summary: Get list of proteins
      tags:
        - assets
  /orders:
    get:
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Order'
                type: array
          description: A list of ingredients
      summary: Get list of orders
      tags:
        - order

  /order:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateOrderResponse'
          description: Order created
      summary: Create an order for existing poke bowls
      tags:
        - order
  /poke:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePokeRequest'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreatePokeResponse'
          description: Poke bowl created
      summary: Create a new poke bowl
      tags:
        - order
  /poke/{id}:
    put:
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePokeRequest'
        required: true
      responses:
        '200':
          description: Poke bowl updated
      summary: Update an existing poke bowl
      tags:
        - order
servers:
  - description: Local development server
    url: http://localhost:8000
tags:
  - description: Endpoints related to assets like bases, proteins, and portions.
    name: assets
  - description: Endpoints related to poke bowls and orders.
    name: order